+++ docklib/__init__.py
@@ -1,5 +1,5 @@
-from .docklib import Dock, DockError
+from .docklib import Dock, DockError, LaunchAgentManager, LaunchAgentError
 
 __version__ = "2.1.0"
 
-__all__ = ["Dock", "DockError"]
+__all__ = ["Dock", "DockError", "LaunchAgentManager", "LaunchAgentError"]